refactor(gui): update status_bar.py import path in window.py and tests
Completes the status_bar.py move from Task 2 -- the file rename landed in
a separate commit due to a stale path in the initial git add.
